requirements.txt
docopt~=0.6.2
lark-parser~=0.9.0
llvmlite~=0.34.0
dataclasses~=0.7; python_version == '3.6'
clang~=11.0